+++ CHANGELOG.md
@@ -10,8 +10,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This is typically useful for boolean parameters. See `Request#hasParameter` ([#49])
 - A basic authentication middleware with pluggable authentication provider ([#28])
 - The request query string. See `Request#query` ([#54])
-- The server host name and the request host name. See `Request#hostname` and `Request#serverHost` ([#55])
-- Access to server port as specified in the host header. See `Request#port` ([#56])
+- The server host name and the request host name, the latter taken from host header. 
+  See `Request#serverHost` and `Request#hostname` ([#55])
+- The server port and the request port, the latter taken from the host header. 
+  See `Request#serverPort` and `Request#port` ([#56])
 - Access to request scheme. See `Request#scheme` ([#58])
  
 ### Changed